Refactoring Widget to also use polymorphism instead of if-else
