+++ src/paymentwall/Pingbacks/PingbackBase.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
-using paymentwall.Pingbacks;
+using Paymentwall.Pingbacks;
 
 namespace Paymentwall.Pingbacks
 {
