+++ src/paymentwall/Pingbacks/VirtualCurrencyPingback.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Paymentwall.Pingbacks;
 
-namespace paymentwall.Pingbacks
+namespace Paymentwall.Pingbacks
 {
     public class VirtualCurrencyPingback : PingbackBase
     {
